+++ testapps/GHIssues/src/main/java/ghissues/Gh819.java
@@ -1,0 +1,31 @@
+package ghissues;
+
+import act.controller.annotation.UrlContext;
+import act.util.JsonView;
+import act.util.LogSupport;
+import org.osgl.mvc.annotation.PostAction;
+
+public abstract class Gh819 extends LogSupport {
+
+    public static class Foo {
+        public String name;
+    }
+
+    @PostAction
+    public Foo test(Foo foo) {
+        return foo;
+    }
+
+    @UrlContext("819")
+    public static class Extended extends Gh819 {
+
+        @JsonView
+        @Override
+        public Foo test(Foo foo) {
+            foo.name = "[extended]" + foo.name;
+            return foo;
+        }
+    }
+
+
+}